refractures functions
more to do...
