+++ planespotting/calculator.py
@@ -1,5 +1,6 @@
 from planespotting.identifiers import *
 from planespotting.utils import *
+from planespotting.decoder import *
 import numpy as np
 
 #All imports end here
